Fixes aad endpoint race condition, tenant switching, and account permissions
